refactor: restructure root configuration and module resolution
